+++ rdsa_utils/io/input.py
@@ -3,12 +3,15 @@ import json
 import logging
 from pathlib import Path
 
-from cloudpathlib import CloudPath
 import tomli
 import yaml
+from cloudpathlib import CloudPath
+from pyspark.sql import DataFrame as SparkDF
+from pyspark.sql import SparkSession
 
+from rdsa_utils.exceptions import DataframeEmptyError
+from rdsa_utils.helpers.pyspark import extract_database_name
 from rdsa_utils.typing import Config
-
 
 logger = logging.getLogger(__name__)
 
@@ -109,3 +112,77 @@ def read_file(file: CloudPath | Path) -> str:
         msg = f'{file=} cannot be found.'
         logger.error(msg)
         raise FileNotFoundError(msg)
+
+
+def load_and_validate_table(
+    spark: SparkSession,
+    table_name: str,
+    skip_validation: bool = False,
+    err_msg: str = None,
+    filter_cond: str = None,
+) -> SparkDF:
+    """Load a table and validate if it is not empty after applying a filter.
+
+    Parameters
+    ----------
+    spark
+        Active SparkSession.
+    table_name
+        Name of the table to load.
+    skip_validation
+        If True, skips validation step, by default False.
+    err_msg
+        Error message to return if table is empty, by default None.
+    filter_cond
+        Condition to apply to SparkDF once read, by default None.
+
+    Returns
+    -------
+    SparkDF
+        Loaded SparkDF if validated, subject to options above.
+
+    Raises
+    ------
+    PermissionError
+        If there's an issue accessing the table or if the table
+        does not exist in the specified database.
+    ValueError
+        If the table is empty after loading, or if it becomes
+        empty after applying a filter condition.
+    """
+    try:
+        df = spark.read.table(table_name)
+        logger.info(f'Successfully loaded table {table_name}.')
+    except Exception as e:
+        db_name, _ = extract_database_name(spark, table_name)
+        db_err = (
+            f'Error accessing {table_name} in the {db_name} database. '
+            'Check you have access to the database and that '
+            'the table name is correct.'
+        )
+        logger.error(db_err)
+        raise PermissionError(db_err) from e
+
+    if not skip_validation:
+        if df.rdd.isEmpty():
+            err_msg = err_msg or f'Table {table_name} is empty.'
+            raise DataframeEmptyError(err_msg)
+
+    if filter_cond:
+        df = df.filter(filter_cond)
+        if not skip_validation and df.rdd.isEmpty():
+            err_msg = (
+                err_msg
+                or f'Table {table_name} is empty after applying '
+                f'filter condition [{filter_cond}].'
+            )
+            raise DataframeEmptyError(err_msg)
+
+    logger.info(
+        (
+            f'Loaded and validated table {table_name}. '
+            f'Filter condition applied: {filter_cond}'
+        ),
+    )
+
+    return df